tests: split into separated executables
